# apps/web/openapi/axon.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Axon REST API",
    "description": "Dedicated REST routes for Axon discovery, RAG, crawl, ingest, and watch workflows.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "6.2.2"
  },
  "paths": {
...
    "/v1/watch": {
      "get": {
        "tags": [
          "watch"
        ],
        "operationId": "list_watch",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Watch definitions",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "Watch storage unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "watch"
        ],
        "operationId": "create_watch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchDefCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created watch definition",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid watch request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "Watch storage unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "The shared, transport-neutral error shape.\n\nEvery crate reports failures as an `ApiError` so CLI, REST, MCP, jobs, logs,\nand progress streams render one consistent structure.",
        "required": [
          "code",
          "message",
          "stage",
          "retryable",
          "severity",
          "visibility",
          "details"
        ],
        "properties": {
          "chunk_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Chunk id."
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable machine code (e.g. `provider.unavailable`)."
          },
          "cooldown_until": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Provider/job cooling timestamp."
          },
          "details": {
            "type": "object",
            "description": "Redacted structured context (safe key/value pairs only).",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "document_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Document id."
          },
          "job_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Job correlation id."
          },
          "message": {
            "type": "string",
            "description": "Redacted human-readable message."
          },
          "provider_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Provider that failed."
          },
          "retry_after_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Suggested retry delay, in milliseconds.",
            "minimum": 0
          },
          "retryable": {
            "type": "boolean",
            "description": "Whether retry may succeed."
          },
          "severity": {
            "$ref": "#/components/schemas/ErrorSeverity",
            "description": "Severity classification."
          },
          "source_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Source id."
          },
          "source_item_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "Item/file/page key."
          },
          "stage": {
            "$ref": "#/components/schemas/ErrorStage",
            "description": "Pipeline/transport stage."
          },
          "visibility": {
            "$ref": "#/components/schemas/ErrorVisibility",
            "description": "Where this error may be surfaced."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "kind",
          "message"
        ],
        "properties": {
          "diagnostics": {
            "type": "object"
          },
          "kind": {
            "$ref": "#/components/schemas/ErrorKind"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "A stable, machine-readable error code (e.g. `provider.unavailable`)."
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "contract_version",
          "error",
          "request_id",
          "trace"
        ],
        "properties": {
          "contract_version": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "ok": {
            "type": "boolean"
          },
          "request_id": {
            "type": "string"
          },
          "trace": {
            "$ref": "#/components/schemas/TraceContext"
          }
        },
        "additionalProperties": false
      },
      "ErrorKind": {
        "type": "string",
        "enum": [
          "bad_gateway",
          "bad_request",
          "challenge_detected",
          "forbidden",
          "internal",
          "invalid_path",
          "ladder_exhausted",
          "not_found",
          "output_dir_error",
          "path_error",
          "path_escape",
          "payload_too_large",
          "rate_limited",
          "read_error",
          "structured_data_malformed",
          "symlink_not_allowed",
          "timeout",
          "unauthorized",
          "unsupported_media_type",
          "upstream_unavailable",
          "vertical_auth_invalid",
          "vertical_auth_missing",
          "vertical_blocked_antibot",
          "vertical_rate_limited",
          "vertical_target_not_found",
          "vertical_target_unavailable",
          "vertical_unsupported_url"
        ]
      },
      "ErrorSeverity": {
        "type": "string",
        "description": "How severe an error is and whether it terminates the item/job.",
        "enum": [
          "info",
          "warning",
          "degraded",
          "failed",
          "fatal"
        ]
      },
      "ErrorStage": {
        "type": "string",
        "description": "Pipeline/transport stage an error is attributed to.",
        "enum": [
          "parsing",
          "validation",
          "resolving",
          "routing",
          "authorizing",
          "planning",
          "leasing",
          "discovering",
          "diffing",
          "fetching",
          "rendering",
          "normalizing",
          "parsing_content",
          "graphing",
          "preparing",
          "embedding",
          "upserting",
          "publishing",
          "cleaning",
          "retrieving",
          "synthesizing",
          "observing"
        ]
      },
      "ErrorVisibility": {
        "type": "string",
        "description": "Where a context entry (or error) may be surfaced.\n\nFrom the \"visibility\" field in the error shape: `public`, `internal`,\n`sensitive`.",
        "enum": [
          "public",
          "internal",
          "sensitive"
        ]
      },
...
      "MetadataMap": {
        "type": "object"
      },
...
      "TraceContext": {
        "type": "object",
        "required": [
          "trace_id",
          "sampled",
          "attributes"
        ],
        "properties": {
          "attributes": {
            "$ref": "#/components/schemas/MetadataMap"
          },
          "parent_span_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "sampled": {
            "type": "boolean"
          },
          "span_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "trace_id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "WatchDefCreateRequest": {
        "type": "object",
        "required": [
          "name",
          "task_type",
          "task_payload",
          "every_seconds"
        ],
        "properties": {
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "every_seconds": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "next_run_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "task_payload": {},
          "task_type": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or static token"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/authorize",
            "tokenUrl": "/token",
            "scopes": {
              "axon:read": "Authenticated Axon REST access",
              "axon:write": "Authenticated Axon REST access"
            }
          }
        }
      }
    }
  },
...
}